File Organization and Name System
Reorganized some of the files and changed their name. F_ in front of the file name means its a function to be called upon. 01, 02, and etc. are all methods to force the main files to be forced to the top of the organized set. Just a visual preference to have the csv files at the bottom.